ProjectEuler+ 001: Use uint64_t instead of int

+++ ProjectEuler/001_multiples.cpp
@@ -1,29 +1,30 @@
+#include <cstdint>
 #include <iostream>
 
-int count(int nMinusOne, int divisor) {
+uint64_t count(uint64_t nMinusOne, int divisor) {
     return nMinusOne/divisor;
 }
 
-int first(int nMinusOne, int divisor) {
+uint64_t first(uint64_t nMinusOne, int divisor) {
     return divisor;
 }
 
-int last(int nMinusOne, int divisor) {
+uint64_t last(uint64_t nMinusOne, int divisor) {
     return count(nMinusOne, divisor) * divisor;
 }
 
-int medianTimesTwo(int nMinusOne, int divisor) {
+uint64_t medianTimesTwo(uint64_t nMinusOne, int divisor) {
     return first(nMinusOne, divisor) + last(nMinusOne, divisor);
 }
 
 // Question: Could one devise a fixed-point number type such
 // that we can leave out this "times two" stuff with zero overhead?
 
-int sumTimesTwo(int nMinusOne, int divisor) {
+uint64_t sumTimesTwo(uint64_t nMinusOne, int divisor) {
     return medianTimesTwo(nMinusOne, divisor) * count(nMinusOne, divisor);
 }
 
-int sumForDivisor(int n, int divisor) {
+uint64_t sumForDivisor(uint64_t n, int divisor) {
     return sumTimesTwo(n-1, divisor) / 2;
 }
 
@@ -33,10 +34,10 @@ int main() {
     std::cin >> t;
 
     for (int i=0; i<t; i++) {
-        int n;
+        uint64_t n;
         std::cin >> n;
 
-        int sum = sumForDivisor(n, 3) + sumForDivisor(n, 5) - sumForDivisor(n, 15);
+        uint64_t sum = sumForDivisor(n, 3) + sumForDivisor(n, 5) - sumForDivisor(n, 15);
         std::cout << sum << std::endl;
     }
